Fix #3 - Add split()
This adds the split() reconciler. It takes a split point, much like
duty(), but instead of taking 2 numbers for the steps, it takes 2
reconcilers and forwards the portions of the phase onto them.

It has only been tested with duty(). It should work with raw() but
almost definitely doesn't work with native().